Cleanup rootless e2e test, simplify logic
