+++ tests/e2e/rootless/Vagrantfile
@@ -1,8 +1,6 @@
 ENV['VAGRANT_NO_PARALLEL'] = 'no'
-NODE_ROLES = (ENV['E2E_NODE_ROLES'] ||
-  ["server-0"])
-NODE_BOXES = (ENV['E2E_NODE_BOXES'] ||
-  ['generic/ubuntu2204'])
+NODE_ROLES = (ENV['E2E_NODE_ROLES'] ||  ["server-0"])
+NODE_BOXES = (ENV['E2E_NODE_BOXES'] ||  ['generic/ubuntu2204'])
 GITHUB_BRANCH = (ENV['E2E_GITHUB_BRANCH'] || "master")
 RELEASE_VERSION = (ENV['E2E_RELEASE_VERSION'] || "")
 GOCOVER = (ENV['E2E_GOCOVER'] || "")
@@ -29,15 +27,16 @@ def provision(vm, role, role_num, node_num)
   vm.provision 'k3s-install', type: 'k3s', run: 'once' do |k3s|
     k3s.args = "server "
     k3s.env = %W[K3S_KUBECONFIG_MODE=0644 INSTALL_K3S_SKIP_START=true INSTALL_K3S_SKIP_ENABLE=true]
+    k3s.env_owner = "vagrant:vagrant"
     k3s.config_mode = '0644' # side-step https://github.com/k3s-io/k3s/issues/4321
+    k3s.config_owner = "vagrant:vagrant"
     k3s.config = <<~YAML
       token: vagrant
       node-external-ip: #{NETWORK_PREFIX}.100
-      flannel-iface: eth1
     YAML
   end
 
-  vm.provision "Enable cgroup v2", type: "shell", path: scripts_location + "/enable_cgroupv2.sh"
+  vm.provision "Download k3s-rootless and enable cgroup v2", type: "shell", path: scripts_location + "/setup_rootless.sh"
   vm.provision 'k3s-reload', type: 'reload', run: 'once'
 end
 
@@ -60,10 +59,7 @@ Vagrant.configure("2") do |config|
     NODE_BOXES = NODE_BOXES.split(" ", -1)
   end
 
-  # Must iterate on the index, vagrant does not understand iterating
-  # over the node roles themselves
-  NODE_ROLES.length.times do |i|
-    name = NODE_ROLES[i]
+  NODE_ROLES.each_with_index do |name, i|
     role_num = name.split("-", -1).pop.to_i
     config.vm.define name do |node|
       provision(node.vm, name, role_num, i)
